Update manual examples, trying to avoid random outputs
Note that doc/example.tex and tst/manual.tst now are out of sync: Some of
these manual examples probably should simply not be turned into tests in the
first place. It doesn't make sense to try to adjust them in the manual right
now, this is something for future revisions.